Split RSSFeed and RSSFeedRegistration
